Quickfix of inputDisplayValue
Up until now inputDisplay showed firstNumber+operator+secondNumber in a
form of 1add2. The reason why as the operator takes its value from the
data-value data attribute, which is a string representation of the operation
, e.g "add".
New variable operatorSign is created and its value equals the last character
of the inputDisplay before any equal sign is pressed.

Example: 1+ is the inputDisplay value, and the second operand stays in the
outputDisplay, so taking the last character is the operator sign.
This is needed to replace 1add2 to the correct arithmetric expression of 1+2.