Refactor number key press logic
Added if statement to track different  key presses when there is an input value (top screen), e.g. when an operator key press lifts the output to the top (from output to input).

Added comments to debug any wrong behaviour.

Note: the two parts of the display might need to be renamed to currentOperandDisplay, previousOperandDisplay.